seatserve-backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
python-decouple==3.8
psycopg2-binary==2.9.9
qrcode==7.4.2
Pillow==10.1.0
django-cors-headers==4.3.1
requests==2.31.0
pytest==7.4.3
pytest-django==4.7.0
stripe==5.9.0
celery==5.3.4
redis==5.0.1

# Production & Deployment
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-redis==5.4.0
python-json-logger==2.0.7

# Error Tracking (optional)
sentry-sdk==1.39.1

# Performance & Monitoring
django-slow-log==0.0.6

